silence/requirements.txt
requests
comtypes
Pillow
pygame
